Activity 10
charts, map, and title change size dynamically to fit screen. Y-axis added to side of chart. Map, top chart, and title show values of specified crop.
Second chart exists and is aligned but is not yet formatted (as this was not part of  Activity 10, but will be used for Lab 2).